third_party/blink/web_tests/platform/generic/fast/forms/enter-clicks-buttons-expected.txt

Test for <rdar://5483519> Pressing Enter on selected buttons should fire onclick





Sending Enter keypresses...




Looping over 6 button elements...


click: BUTTON type=button in form
click: BUTTON type=submit in form
submit: FORM
click: BUTTON type=reset in form
click: BUTTON type=button
click: BUTTON type=submit
click: BUTTON type=reset


Looping over 8 input elements...


click: BUTTON type=submit in form
submit: FORM
click: INPUT type=button in form
click: INPUT type=submit in form
submit: FORM
click: INPUT type=reset in form
click: INPUT type=button
click: INPUT type=submit
click: INPUT type=reset




Sending U+0020 keypresses...




Looping over 6 button elements...


click: BUTTON type=button in form
click: BUTTON type=submit in form
submit: FORM
click: BUTTON type=reset in form
click: BUTTON type=button
click: BUTTON type=submit
click: BUTTON type=reset


Looping over 8 input elements...


click: INPUT type=checkbox in form
click: INPUT type=button in form
click: INPUT type=submit in form
submit: FORM
click: INPUT type=reset in form
click: INPUT type=checkbox
click: INPUT type=button
click: INPUT type=submit
click: INPUT type=reset
PASS successfullyParsed is true

TEST COMPLETE
